requirements.txt
Django==1.6.2
Pillow==2.3.1
Pygments==1.6
South==0.8.4
bpython==0.12
django-haystack==2.1.0
elasticsearch==1.0.0
psycopg2==2.5.2
pyelasticsearch==0.6.1
requests==2.2.1
simplejson==3.3.3
six==1.6.1
urllib3==1.8
virtualenv==1.11.4
